incorporate FluxTable into pfsObject and pfsSingle
This requires some changes to the plumbing, because we want to build
the FluxTable from unresampled arm spectra, but we were previously
building the coadds from the merged-arm spectra for convenience.

Made FocalPlaneFunction care about wavelength (like SubtractSky1DSolution,
so we no longer need that) so we can use it to correct both resampled and
unresampled spectra. Added methods to MeasureFluxCalibrationTask and
SubtractSky1DTask (which use FocalPlaneFunction) to apply their corrections
to either PfsSpectra (e.g., a pfsArm) or PfsSpectrum (e.g., pfsSingle).
Changed CoaddSpectraTask so that it applies the flux calibration and
1D sky subtraction to individual arm spectra instead of doing the merge.

+++ python/pfs/drp/stella/subtractSky2d.py
@@ -1,3 +1,4 @@
+import numpy as np
 from lsst.pex.config import Config, ConfigurableField
 from lsst.pipe.base import Task
 
@@ -106,7 +107,7 @@ class SubtractSky2dTask(Task):
             spectrum.setWavelength(detectorMap.getWavelength(ft.fiberId))
         fluxes = sky2d(spectra.getAllWavelengths(), centers)
         for ss, flux in zip(spectra, fluxes):
-            ss.spectrum = flux
+            ss.spectrum = flux.astype(np.float32)
 
         skyImage = spectra.makeImage(exposure.getBBox(), fiberTrace)
         exposure.maskedImage -= skyImage
